requirement.txt
# Core dependencies
pandas==1.3.3
tqdm==4.62.3
Pillow==8.3.2
easyocr==1.4.1
torch==1.7.1+cu101  # or compatible with your CUDA setup (update if needed)
spacy==3.1.3
langdetect==1.0.9
# Additional requirements for spaCy models